requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-bootstrap5==22.2
django-filter==21.1
django-nested-admin==4.0.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.0
flake8==6.0.0
flake8-isort==6.0.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
Pillow==9.3.0
pluggy==1.4.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.4.0
python-dotenv==1.0.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
